requirements.txt
streamlit
pdfminer.six
python-docx
pandas
scikit-learn
numpy
